# [Multithread] Step 3: Parallel TableScan