--- a/src/codegen/operator/delete_translator.cpp
+++ b/src/codegen/operator/delete_translator.cpp
@@ -40,11 +40,10 @@ void DeleteTranslator::InitializeState() {
 
   // Get the table pointer
   storage::DataTable *table = delete_plan_.GetTable();
-  llvm::Value *table_ptr =
-      codegen.Call(StorageManagerProxy::GetTableWithOid,
-                   {GetStorageManagerPtr(),
-                    codegen.Const32(table->GetDatabaseOid()),
-                    codegen.Const32(table->GetOid())});
+  llvm::Value *table_ptr = codegen.Call(
+      StorageManagerProxy::GetTableWithOid,
+      {GetStorageManagerPtr(), codegen.Const32(table->GetDatabaseOid()),
+       codegen.Const32(table->GetOid())});
 
   llvm::Value *executor_ptr = GetCompilationContext().GetExecutorContextPtr();