--- a/src/codegen/operator/delete_translator.cpp
+++ b/src/codegen/operator/delete_translator.cpp
@@ -53,9 +53,9 @@ void DeleteTranslator::InitializeState() {
   codegen.Call(DeleterProxy::Init, {deleter, table_ptr, executor_ptr});
 }
 
-void DeleteTranslator::Produce() const {
+std::vector<CodeGenStage> DeleteTranslator::Produce() const {
   // Call Produce() on our child (a scan), to produce the tuples we'll delete
-  GetCompilationContext().Produce(*delete_plan_.GetChild(0));
+  return GetCompilationContext().Produce(*delete_plan_.GetChild(0));
 }
 
 void DeleteTranslator::Consume(ConsumerContext &, RowBatch::Row &row) const {