… src/codegen/operator/hash_group_by_translator.cpp
… src/codegen/operator/hash_join_translator.cpp
… src/codegen/operator/insert_translator.cpp
… src/codegen/operator/order_by_translator.cpp
… src/codegen/operator/projection_translator.cpp
… src/codegen/operator/table_scan_translator.cpp
… src/codegen/operator/update_translator.cpp
… src/codegen/pipeline.cpp
… src/codegen/query.cpp
… src/codegen/runtime_functions.cpp
… src/codegen/table.cpp
… src/codegen/tile_group.cpp
… src/codegen/transaction_runtime.cpp
… src/concurrency/transaction_context.cpp
… src/executor/plan_executor.cpp
… src/expression/abstract_expression.cpp
… src/include/codegen/auxiliary_producer_function.h
… src/include/codegen/buffering_consumer.h
… src/include/codegen/compilation_context.h
… src/include/codegen/consumer_context.h
… src/include/codegen/counting_consumer.h
… src/include/codegen/operator/block_nested_loop_join_translator.h
… src/include/codegen/operator/delete_translator.h
… src/include/codegen/operator/global_group_by_translator.h
… src/include/codegen/operator/hash_group_by_translator.h
… src/include/codegen/operator/hash_join_translator.h
… src/include/codegen/operator/hash_translator.h
… src/include/codegen/operator/insert_translator.h
… src/include/codegen/operator/operator_translator.h
… src/include/codegen/operator/order_by_translator.h