- Each consumer is notified the number of tasks, so that it can prepare one result vector for each task.
- Each task pushes to a separate vector.
- <del>RWSet was `oid -> oid -> RWType`, now is `ItemPointer -> RWType` and uses cuckoo hash map.</del>