@seojungmin 2018-01-10
@phisiart Can you run a brief scalability test?
- 1 thread (before, i.e.without this PR) vs. 1 thread(now, i.e. with this PR)
- 2,4,8 threads (with this PR)