@phisiart 2018-01-11
@seojungmin 

I tested scalability of this PR on my own laptop (4 cores 8 hyperthreads).

I'll compare before & after this PR later.

Setup:
1,000,000 tuples;
predicate: a >= ? and b >= a such that selectivity = 0.5;
run each test 5 times;
ntuples / tilegroup = 1000;
ntasks = nthreads in monoqueuepool.

```
1 Task:

Test table has 1000 tile groups
CONFIGURATION:
===============
Scan complexity: 1, Selectivity: 0.50
2018-01-10 22:57:13 [test/codegen/scan_benchmark_test.cpp:166:RunCompiledExperiment] INFO  - plan time: 585.007440
2018-01-10 22:57:13 [test/codegen/scan_benchmark_test.cpp:166:RunCompiledExperiment] INFO  - plan time: 551.020922
2018-01-10 22:57:14 [test/codegen/scan_benchmark_test.cpp:166:RunCompiledExperiment] INFO  - plan time: 526.301425
2018-01-10 22:57:15 [test/codegen/scan_benchmark_test.cpp:166:RunCompiledExperiment] INFO  - plan time: 597.413579
2018-01-10 22:57:15 [test/codegen/scan_benchmark_test.cpp:166:RunCompiledExperiment] INFO  - plan time: 624.516938

2 Tasks:

Test table has 1000 tile groups
CONFIGURATION:
===============
Scan complexity: 1, Selectivity: 0.50
2018-01-10 22:58:00 [test/codegen/scan_benchmark_test.cpp:166:RunCompiledExperiment] INFO  - plan time: 320.773205
2018-01-10 22:58:01 [test/codegen/scan_benchmark_test.cpp:166:RunCompiledExperiment] INFO  - plan time: 348.360585
2018-01-10 22:58:01 [test/codegen/scan_benchmark_test.cpp:166:RunCompiledExperiment] INFO  - plan time: 380.253163
2018-01-10 22:58:02 [test/codegen/scan_benchmark_test.cpp:166:RunCompiledExperiment] INFO  - plan time: 363.823048
2018-01-10 22:58:02 [test/codegen/scan_benchmark_test.cpp:166:RunCompiledExperiment] INFO  - plan time: 331.432027

4 Tasks:

Test table has 1000 tile groups
CONFIGURATION:
===============
Scan complexity: 1, Selectivity: 0.50
2018-01-10 22:59:22 [test/codegen/scan_benchmark_test.cpp:166:RunCompiledExperiment] INFO  - plan time: 246.302288
2018-01-10 22:59:22 [test/codegen/scan_benchmark_test.cpp:166:RunCompiledExperiment] INFO  - plan time: 241.638012
2018-01-10 22:59:23 [test/codegen/scan_benchmark_test.cpp:166:RunCompiledExperiment] INFO  - plan time: 230.019421
2018-01-10 22:59:23 [test/codegen/scan_benchmark_test.cpp:166:RunCompiledExperiment] INFO  - plan time: 248.375334
2018-01-10 22:59:23 [test/codegen/scan_benchmark_test.cpp:166:RunCompiledExperiment] INFO  - plan time: 230.579879

8 Tasks:

Test table has 1000 tile groups
CONFIGURATION:
===============
Scan complexity: 1, Selectivity: 0.50
2018-01-10 23:00:09 [test/codegen/scan_benchmark_test.cpp:166:RunCompiledExperiment] INFO  - plan time: 328.316249
2018-01-10 23:00:10 [test/codegen/scan_benchmark_test.cpp:166:RunCompiledExperiment] INFO  - plan time: 306.291659
2018-01-10 23:00:10 [test/codegen/scan_benchmark_test.cpp:166:RunCompiledExperiment] INFO  - plan time: 292.808897
2018-01-10 23:00:10 [test/codegen/scan_benchmark_test.cpp:166:RunCompiledExperiment] INFO  - plan time: 293.221811
2018-01-10 23:00:11 [test/codegen/scan_benchmark_test.cpp:166:RunCompiledExperiment] INFO  - plan time: 248.828105
```

If we comment out PerformRead:

```
1 Task (No PerformRead):

Test table has 1000 tile groups
CONFIGURATION:
===============
Scan complexity: 1, Selectivity: 0.50
2018-01-10 22:54:02 [test/codegen/scan_benchmark_test.cpp:166:RunCompiledExperiment] INFO  - plan time: 141.522882
2018-01-10 22:54:02 [test/codegen/scan_benchmark_test.cpp:166:RunCompiledExperiment] INFO  - plan time: 146.596603
2018-01-10 22:54:02 [test/codegen/scan_benchmark_test.cpp:166:RunCompiledExperiment] INFO  - plan time: 151.640471
2018-01-10 22:54:03 [test/codegen/scan_benchmark_test.cpp:166:RunCompiledExperiment] INFO  - plan time: 134.416357
2018-01-10 22:54:03 [test/codegen/scan_benchmark_test.cpp:166:RunCompiledExperiment] INFO  - plan time: 146.240653

2 Tasks (No PerformRead):

Test table has 1000 tile groups
CONFIGURATION:
===============
Scan complexity: 1, Selectivity: 0.50
2018-01-10 22:54:57 [test/codegen/scan_benchmark_test.cpp:166:RunCompiledExperiment] INFO  - plan time: 109.912670
2018-01-10 22:54:57 [test/codegen/scan_benchmark_test.cpp:166:RunCompiledExperiment] INFO  - plan time: 114.105909
2018-01-10 22:54:57 [test/codegen/scan_benchmark_test.cpp:166:RunCompiledExperiment] INFO  - plan time: 102.025413
2018-01-10 22:54:57 [test/codegen/scan_benchmark_test.cpp:166:RunCompiledExperiment] INFO  - plan time: 136.733364
2018-01-10 22:54:58 [test/codegen/scan_benchmark_test.cpp:166:RunCompiledExperiment] INFO  - plan time: 142.408028

4 Tasks (No PerformRead):

Test table has 1000 tile groups
CONFIGURATION:
===============
Scan complexity: 1, Selectivity: 0.50
2018-01-10 22:52:16 [test/codegen/scan_benchmark_test.cpp:166:RunCompiledExperiment] INFO  - plan time: 91.179798
2018-01-10 22:52:16 [test/codegen/scan_benchmark_test.cpp:166:RunCompiledExperiment] INFO  - plan time: 47.328278
2018-01-10 22:52:16 [test/codegen/scan_benchmark_test.cpp:166:RunCompiledExperiment] INFO  - plan time: 44.501548
2018-01-10 22:52:16 [test/codegen/scan_benchmark_test.cpp:166:RunCompiledExperiment] INFO  - plan time: 46.975591
2018-01-10 22:52:16 [test/codegen/scan_benchmark_test.cpp:166:RunCompiledExperiment] INFO  - plan time: 43.201461

8 Tasks (No PerformRead):

Test table has 1000 tile groups
CONFIGURATION:
===============
Scan complexity: 1, Selectivity: 0.50
2018-01-10 22:51:06 [test/codegen/scan_benchmark_test.cpp:166:RunCompiledExperiment] INFO  - plan time: 34.236184
2018-01-10 22:51:06 [test/codegen/scan_benchmark_test.cpp:166:RunCompiledExperiment] INFO  - plan time: 37.056811
2018-01-10 22:51:06 [test/codegen/scan_benchmark_test.cpp:166:RunCompiledExperiment] INFO  - plan time: 33.332180
2018-01-10 22:51:06 [test/codegen/scan_benchmark_test.cpp:166:RunCompiledExperiment] INFO  - plan time: 34.758514
2018-01-10 22:51:06 [test/codegen/scan_benchmark_test.cpp:166:RunCompiledExperiment] INFO  - plan time: 34.382642
```