@phisiart 2018-01-11
New setup:
1,000,000 tuples;
predicate: a >= ? and b >= a such that selectivity = 0.5;
run each test 5 times;
ntuples / tilegroup = 125,000 (so 8 tilegroups);
ntasks = nthreads in monoqueuepool.

This time I also did the comparison between before & after PR.

Results:
In this PR, by making rwset thread-safe and using MonoQueuePool, performance drops significantly: for 1 thread, running time = 2x (221ms -> 436ms, which means 200ms of overhead!).
Scalability stuck at ~ 2x.

By turning off visibility check and PerformRead (i.e. not modifying rwset), performance is still 2x for 1 thread (59 ms -> 117 ms, which means 50ms of overhead!). Scalability gets to ~4x (which matches my CPU's number of cores?)

An interesting observation is that if I turn off exponential backoff in MonoQueuePool, the overhead drops to 15ms.

```
==================================================================================================================
Experiment: Before/after PR comparison, scalability.

1 Task: (Time = 1x)

Test table has 8 tile groups
CONFIGURATION:
===============
Scan complexity: 1, Selectivity: 0.50
2018-01-11 14:41:31 [test/codegen/scan_benchmark_test.cpp:167:RunCompiledExperiment] INFO  - plan time: 435.952909
2018-01-11 14:41:32 [test/codegen/scan_benchmark_test.cpp:167:RunCompiledExperiment] INFO  - plan time: 514.468161
2018-01-11 14:41:32 [test/codegen/scan_benchmark_test.cpp:167:RunCompiledExperiment] INFO  - plan time: 473.257402
2018-01-11 14:41:33 [test/codegen/scan_benchmark_test.cpp:167:RunCompiledExperiment] INFO  - plan time: 506.756334
2018-01-11 14:41:33 [test/codegen/scan_benchmark_test.cpp:167:RunCompiledExperiment] INFO  - plan time: 517.355382

1 Task (before PR): (Time = 0.51x)

Test table has 8 tile groups
CONFIGURATION:
===============
Scan complexity: 1, Selectivity: 0.50
2018-01-11 15:05:23 [test/codegen/scan_benchmark_test.cpp:167:RunCompiledExperiment] INFO  - plan time: 230.127799
2018-01-11 15:05:23 [test/codegen/scan_benchmark_test.cpp:167:RunCompiledExperiment] INFO  - plan time: 223.853070
2018-01-11 15:05:23 [test/codegen/scan_benchmark_test.cpp:167:RunCompiledExperiment] INFO  - plan time: 221.554144
2018-01-11 15:05:24 [test/codegen/scan_benchmark_test.cpp:167:RunCompiledExperiment] INFO  - plan time: 222.420255
2018-01-11 15:05:24 [test/codegen/scan_benchmark_test.cpp:167:RunCompiledExperiment] INFO  - plan time: 222.193789

2 Tasks: (Time = 0.61x)

Test table has 8 tile groups
CONFIGURATION:
===============
Scan complexity: 1, Selectivity: 0.50
2018-01-11 14:42:16 [test/codegen/scan_benchmark_test.cpp:167:RunCompiledExperiment] INFO  - plan time: 269.678217
2018-01-11 14:42:17 [test/codegen/scan_benchmark_test.cpp:167:RunCompiledExperiment] INFO  - plan time: 267.280748
2018-01-11 14:42:17 [test/codegen/scan_benchmark_test.cpp:167:RunCompiledExperiment] INFO  - plan time: 266.876594
2018-01-11 14:42:17 [test/codegen/scan_benchmark_test.cpp:167:RunCompiledExperiment] INFO  - plan time: 268.253595
2018-01-11 14:42:18 [test/codegen/scan_benchmark_test.cpp:167:RunCompiledExperiment] INFO  - plan time: 284.334824

4 Tasks: (Time = 0.39x)

Test table has 8 tile groups
CONFIGURATION:
===============
Scan complexity: 1, Selectivity: 0.50
2018-01-11 14:43:26 [test/codegen/scan_benchmark_test.cpp:167:RunCompiledExperiment] INFO  - plan time: 218.361114
2018-01-11 14:43:26 [test/codegen/scan_benchmark_test.cpp:167:RunCompiledExperiment] INFO  - plan time: 188.463938
2018-01-11 14:43:27 [test/codegen/scan_benchmark_test.cpp:167:RunCompiledExperiment] INFO  - plan time: 180.831867
2018-01-11 14:43:27 [test/codegen/scan_benchmark_test.cpp:167:RunCompiledExperiment] INFO  - plan time: 181.302950
2018-01-11 14:43:27 [test/codegen/scan_benchmark_test.cpp:167:RunCompiledExperiment] INFO  - plan time: 171.998242

8 Tasks: (Time = 0.45x)

Test table has 8 tile groups
CONFIGURATION:
===============
Scan complexity: 1, Selectivity: 0.50
2018-01-11 14:44:16 [test/codegen/scan_benchmark_test.cpp:167:RunCompiledExperiment] INFO  - plan time: 243.041674
2018-01-11 14:44:16 [test/codegen/scan_benchmark_test.cpp:167:RunCompiledExperiment] INFO  - plan time: 221.510711
2018-01-11 14:44:16 [test/codegen/scan_benchmark_test.cpp:167:RunCompiledExperiment] INFO  - plan time: 220.037998
2018-01-11 14:44:17 [test/codegen/scan_benchmark_test.cpp:167:RunCompiledExperiment] INFO  - plan time: 209.665845
2018-01-11 14:44:17 [test/codegen/scan_benchmark_test.cpp:167:RunCompiledExperiment] INFO  - plan time: 195.996666

==================================================================================================================
Experiment: Turning off PerformRead (no update to rw_set)

1 Task (No PerformRead): (Time = 1x)

Test table has 8 tile groups
CONFIGURATION:
===============
Scan complexity: 1, Selectivity: 0.50
2018-01-11 14:47:44 [test/codegen/scan_benchmark_test.cpp:167:RunCompiledExperiment] INFO  - plan time: 93.553732
2018-01-11 14:47:44 [test/codegen/scan_benchmark_test.cpp:167:RunCompiledExperiment] INFO  - plan time: 109.886442
2018-01-11 14:47:44 [test/codegen/scan_benchmark_test.cpp:167:RunCompiledExperiment] INFO  - plan time: 120.619925
2018-01-11 14:47:44 [test/codegen/scan_benchmark_test.cpp:167:RunCompiledExperiment] INFO  - plan time: 149.485346
2018-01-11 14:47:45 [test/codegen/scan_benchmark_test.cpp:167:RunCompiledExperiment] INFO  - plan time: 150.201550

1 Task (No PerformRead, before PR): (Time = 0.71x)

Test table has 8 tile groups
CONFIGURATION:
===============
Scan complexity: 1, Selectivity: 0.50
2018-01-11 15:11:57 [test/codegen/scan_benchmark_test.cpp:167:RunCompiledExperiment] INFO  - plan time: 66.765122
2018-01-11 15:11:57 [test/codegen/scan_benchmark_test.cpp:167:RunCompiledExperiment] INFO  - plan time: 67.246447
2018-01-11 15:11:57 [test/codegen/scan_benchmark_test.cpp:167:RunCompiledExperiment] INFO  - plan time: 71.626163
2018-01-11 15:11:57 [test/codegen/scan_benchmark_test.cpp:167:RunCompiledExperiment] INFO  - plan time: 70.611501
2018-01-11 15:11:57 [test/codegen/scan_benchmark_test.cpp:167:RunCompiledExperiment] INFO  - plan time: 70.415887

2 Tasks (No PerformRead): (Time = 0.88x)

Test table has 8 tile groups
CONFIGURATION:
===============
Scan complexity: 1, Selectivity: 0.50
2018-01-11 14:46:46 [test/codegen/scan_benchmark_test.cpp:167:RunCompiledExperiment] INFO  - plan time: 83.586065
2018-01-11 14:46:46 [test/codegen/scan_benchmark_test.cpp:167:RunCompiledExperiment] INFO  - plan time: 82.452345
2018-01-11 14:46:46 [test/codegen/scan_benchmark_test.cpp:167:RunCompiledExperiment] INFO  - plan time: 104.965656
2018-01-11 14:46:46 [test/codegen/scan_benchmark_test.cpp:167:RunCompiledExperiment] INFO  - plan time: 122.975478
2018-01-11 14:46:46 [test/codegen/scan_benchmark_test.cpp:167:RunCompiledExperiment] INFO  - plan time: 109.412292

4 Tasks (No PerformRead): (Time = 0.51x)

Test table has 8 tile groups
CONFIGURATION:
===============
Scan complexity: 1, Selectivity: 0.50
2018-01-11 14:45:52 [test/codegen/scan_benchmark_test.cpp:167:RunCompiledExperiment] INFO  - plan time: 47.989170
2018-01-11 14:45:52 [test/codegen/scan_benchmark_test.cpp:167:RunCompiledExperiment] INFO  - plan time: 66.631455
2018-01-11 14:45:53 [test/codegen/scan_benchmark_test.cpp:167:RunCompiledExperiment] INFO  - plan time: 47.803405
2018-01-11 14:45:53 [test/codegen/scan_benchmark_test.cpp:167:RunCompiledExperiment] INFO  - plan time: 66.911422
2018-01-11 14:45:53 [test/codegen/scan_benchmark_test.cpp:167:RunCompiledExperiment] INFO  - plan time: 47.799438

8 Tasks (No PerformRead): (Time = 0.37x)

Test table has 8 tile groups
CONFIGURATION:
===============
Scan complexity: 1, Selectivity: 0.50
2018-01-11 14:44:58 [test/codegen/scan_benchmark_test.cpp:167:RunCompiledExperiment] INFO  - plan time: 38.233067
2018-01-11 14:44:58 [test/codegen/scan_benchmark_test.cpp:167:RunCompiledExperiment] INFO  - plan time: 39.846321
2018-01-11 14:44:58 [test/codegen/scan_benchmark_test.cpp:167:RunCompiledExperiment] INFO  - plan time: 36.961351
2018-01-11 14:44:58 [test/codegen/scan_benchmark_test.cpp:167:RunCompiledExperiment] INFO  - plan time: 35.404664
2018-01-11 14:44:58 [test/codegen/scan_benchmark_test.cpp:167:RunCompiledExperiment] INFO  - plan time: 35.044512

==================================================================================================================
Experiment: No visibility check, no PerformRead.

1 Task (No IsVisible, No PerformRead): (Time = 1x)

Test table has 8 tile groups
CONFIGURATION:
===============
Scan complexity: 1, Selectivity: 0.50
2018-01-11 14:51:08 [test/codegen/scan_benchmark_test.cpp:167:RunCompiledExperiment] INFO  - plan time: 120.064447
2018-01-11 14:51:08 [test/codegen/scan_benchmark_test.cpp:167:RunCompiledExperiment] INFO  - plan time: 139.241952
2018-01-11 14:51:08 [test/codegen/scan_benchmark_test.cpp:167:RunCompiledExperiment] INFO  - plan time: 117.924046
2018-01-11 14:51:08 [test/codegen/scan_benchmark_test.cpp:167:RunCompiledExperiment] INFO  - plan time: 132.782871
2018-01-11 14:51:09 [test/codegen/scan_benchmark_test.cpp:167:RunCompiledExperiment] INFO  - plan time: 138.358248

1 Task (No IsVisible, No PerformRead, before PR): (Time = 0.5x)

Test table has 8 tile groups
CONFIGURATION:
===============
Scan complexity: 1, Selectivity: 0.50
2018-01-11 15:12:53 [test/codegen/scan_benchmark_test.cpp:167:RunCompiledExperiment] INFO  - plan time: 59.118067
2018-01-11 15:12:53 [test/codegen/scan_benchmark_test.cpp:167:RunCompiledExperiment] INFO  - plan time: 60.068787
2018-01-11 15:12:53 [test/codegen/scan_benchmark_test.cpp:167:RunCompiledExperiment] INFO  - plan time: 62.281756
2018-01-11 15:12:53 [test/codegen/scan_benchmark_test.cpp:167:RunCompiledExperiment] INFO  - plan time: 61.056381
2018-01-11 15:12:53 [test/codegen/scan_benchmark_test.cpp:167:RunCompiledExperiment] INFO  - plan time: 61.123222

1 Task (No IsVisible, No PerformRead, No exponential backoff in MonoQueuePool): (Time = 0.63x)

Test table has 8 tile groups
CONFIGURATION:
===============
Scan complexity: 1, Selectivity: 0.50
2018-01-11 15:45:50 [test/codegen/scan_benchmark_test.cpp:167:RunCompiledExperiment] INFO  - plan time: 79.615601
2018-01-11 15:45:50 [test/codegen/scan_benchmark_test.cpp:167:RunCompiledExperiment] INFO  - plan time: 74.528722
2018-01-11 15:45:50 [test/codegen/scan_benchmark_test.cpp:167:RunCompiledExperiment] INFO  - plan time: 74.354281
2018-01-11 15:45:50 [test/codegen/scan_benchmark_test.cpp:167:RunCompiledExperiment] INFO  - plan time: 74.724083
2018-01-11 15:45:50 [test/codegen/scan_benchmark_test.cpp:167:RunCompiledExperiment] INFO  - plan time: 74.643463

2 Tasks (No IsVisible, No PerformRead): (Time = 0.71x)

Test table has 8 tile groups
CONFIGURATION:
===============
Scan complexity: 1, Selectivity: 0.50
2018-01-11 14:53:22 [test/codegen/scan_benchmark_test.cpp:167:RunCompiledExperiment] INFO  - plan time: 84.660740
2018-01-11 14:53:23 [test/codegen/scan_benchmark_test.cpp:167:RunCompiledExperiment] INFO  - plan time: 96.440224
2018-01-11 14:53:23 [test/codegen/scan_benchmark_test.cpp:167:RunCompiledExperiment] INFO  - plan time: 92.467271
2018-01-11 14:53:23 [test/codegen/scan_benchmark_test.cpp:167:RunCompiledExperiment] INFO  - plan time: 123.923090
2018-01-11 14:53:23 [test/codegen/scan_benchmark_test.cpp:167:RunCompiledExperiment] INFO  - plan time: 135.806605

4 Tasks (No IsVisible, No PerformRead): (Time = 0.32x)

Test table has 8 tile groups
CONFIGURATION:
===============
Scan complexity: 1, Selectivity: 0.50
2018-01-11 14:55:11 [test/codegen/scan_benchmark_test.cpp:167:RunCompiledExperiment] INFO  - plan time: 51.815064
2018-01-11 14:55:12 [test/codegen/scan_benchmark_test.cpp:167:RunCompiledExperiment] INFO  - plan time: 52.762700
2018-01-11 14:55:12 [test/codegen/scan_benchmark_test.cpp:167:RunCompiledExperiment] INFO  - plan time: 44.049860
2018-01-11 14:55:12 [test/codegen/scan_benchmark_test.cpp:167:RunCompiledExperiment] INFO  - plan time: 38.630543
2018-01-11 14:55:12 [test/codegen/scan_benchmark_test.cpp:167:RunCompiledExperiment] INFO  - plan time: 43.808900

8 Tasks (No IsVisible, No PerformRead): (Time = 0.23x)

Test table has 8 tile groups
CONFIGURATION:
===============
Scan complexity: 1, Selectivity: 0.50
2018-01-11 14:56:09 [test/codegen/scan_benchmark_test.cpp:167:RunCompiledExperiment] INFO  - plan time: 38.013033
2018-01-11 14:56:09 [test/codegen/scan_benchmark_test.cpp:167:RunCompiledExperiment] INFO  - plan time: 28.162359
2018-01-11 14:56:09 [test/codegen/scan_benchmark_test.cpp:167:RunCompiledExperiment] INFO  - plan time: 33.620135
2018-01-11 14:56:09 [test/codegen/scan_benchmark_test.cpp:167:RunCompiledExperiment] INFO  - plan time: 31.392653
2018-01-11 14:56:09 [test/codegen/scan_benchmark_test.cpp:167:RunCompiledExperiment] INFO  - plan time: 31.461934
```