@seojungmin 2018-01-12
For your information, this was a brief *old* multithread version benchmark result which I ran on dev4 a while ago.  The visibility-check performance may differ in this version, due to the different read-write set algorithm, but you can see the visibility checking overhead.

Without numa setting, (dev4)
1 thread: 201~215ms
2 threads: 90~137ms
4 threads: 61-72ms
8 threads: 33~39ms

With all threads set to a CPU socket,
1 thread: 152~164ms
2 threads: 69-72ms
4 threads: 46~54ms
8 threads: 29~30ms
16 threads: 20ms 

Without the visibility check,
1 thread: ~57ms
4 threads:~19ms
8 threads; 10ms
16 threads: 7ms